using jqgrid

+++ EntityLayer/Concrete/MusteriUrunDto.cs
@@ -11,7 +11,7 @@ namespace EntityLayer.Concrete
         public int UrunId { get; set; }
         public string UrunAdi { get; set; }
 
-        public int UrunFiyati { get; set; }
+        public decimal UrunFiyati { get; set; }
 
         public int UrunStok { get; set; }
     }
